protect live view with feature flag
