+++ src/ui/ViewModels/FromUcasViewModel.cs
@@ -7,6 +7,7 @@ namespace GovUk.Education.ManageCourses.Ui.ViewModels
     {
         public bool AllowPreview { get; set; }
         public bool AllowPublish { get;  set; }
+        public bool AllowLiveView { get;  set; }
         public string OrganisationName { get; set; }
         public string OrganisationId { get; set; }
         public bool MultipleOrganisations { get; set; }
